+++ IMP/IMP.Application/Models/ViewModels/PageViewModel.cs
@@ -9,7 +9,7 @@ namespace IMP.Application.Models.ViewModels
     public class PageViewModel : BaseViewModel<int>
     {
         public string BackgroundType { get; set; }
-        public int FontSize { get; set; }
+        public double FontSize { get; set; }
         public string Background { get; set; }
         public string BioLink { get; set; }
         public string FontFamily { get; set; }
